Refactor sesh
