+++ client/components/forms/user_input_form.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import NavBar from '../ui/nav';
-import axios from 'axios';
-import async from 'async';
 import UserRideSelection from '../user/user_ride_selection';
 import * as MapTools from '../../util/cartographic_tools';
 import * as LatLongTool from '../../util/latlong_conversion';
-
+import * as AlgorithmLogic from '../../util/algorithm_logic';
 
 class UserInputForm extends React.Component {
   constructor(props) {
@@ -21,15 +19,15 @@ class UserInputForm extends React.Component {
       rideType: "lyft",
     };
 
-    this.updateAddress = this.updateAddress.bind(this);
     this.updateInput = this.updateInput.bind(this);
     this.submitForm = this.submitForm.bind(this);
-    this.rideEstimate = this.rideEstimate.bind(this);
-    this.getBoundaries = this.getBoundaries.bind(this);
     this.centerMap = this.centerMap.bind(this);
     this.getRideType = this.getRideType.bind(this);
 
     this.parseAddressToLatLng = LatLongTool.parseAddressToLatLng.bind(this);
+    this.getBoundaries = AlgorithmLogic.getBoundaries.bind(this);
+    this.landOrWater = AlgorithmLogic.landOrWater.bind(this);
+    this.rideEstimate = AlgorithmLogic.rideEstimate.bind(this);
   }
 
   submitForm(e) {
@@ -49,120 +47,8 @@ class UserInputForm extends React.Component {
     return (e) => { this.setState({ [field]: e.target.value }); };
   }
 
-  updateAddress(e) {
-    this.setState({ addressInput: e.target.value });
-  }
-
-//TODO: Refactoring
-
-  // parseAddressToLatLng(address, callback) {
-  //  const geocoder = new google.maps.Geocoder;
-  //  geocoder.geocode({ address: address }, (results, status) => {
-  //    if (status === 'OK') {
-  //      const addressLatLng = new google.maps.LatLng(
-  //        results[0].geometry.location.lat(),
-  //        results[0].geometry.location.lng()
-  //      );
-  //      this.setState({ addressLatLng }, () => {
-  //        this.getBoundaries();
-  //      });
-  //      this.centerMap(addressLatLng);
-  //    }
-  //  });
-  // }
-
-  getBoundaries() {
-    const stdDev = 2;
-    const amount = 15;
-    const defaultRadiusInMeters = 32000;
-    const currentLatLng = this.state.addressLatLng;
-    let directions = [];
-
-    for (let i = 0; i < 360; i+=20) {
-      directions.push(i);
-    }
-
-    const googleGeometry = google.maps.geometry.spherical;
-
-    async.eachOf(directions, (direction, index, callback) => {
-      const endLatLng = new googleGeometry.computeOffset(currentLatLng, defaultRadiusInMeters, direction);
-      // this.landOrWater(endLatLng.lat(), endLatLng.lng(), res => console.log(res))
-      this.rideEstimate(currentLatLng, endLatLng, amount, stdDev, index, direction, []);
-      callback(null);
-    });
-  }
-
   getRideType(type) {
-    this.setState({ rideType: type }, () => { this.getBoundaries() })
-  }
-
-  landOrWater(lat, lng, callback) {
-      const map_url = "http://maps.googleapis.com/maps/api/staticmap?center="+lat+","+lng+"&zoom="+this.props.map.getZoom()+"&size=1x1&maptype=roadmap"
-      const canvas = document.createElement('canvas');
-      const ctx = canvas.getContext('2d');
-      let result;
-
-      const image = new Image();
-      image.crossOrigin = "Anonymous";
-      image.src = map_url;
-
-      image.onload = function() {
-          canvas.width = image.width;
-          canvas.height = image.height;
-          canvas.getContext('2d').drawImage(image, 0, 0, image.width, image.height);
-          const pixelData = canvas.getContext('2d').getImageData(0, 0, 1, 1).data;
-          if( pixelData[0] > 160 && pixelData[0] < 181 && pixelData[1] > 190 && pixelData[1] < 210 ) {
-              result = "water";
-          } else {
-              result = "land";
-          }
-          callback(result);
-      }
-  }
-
-  async rideEstimate(start, end, amount, stdDev, index, direction, history) {
-    let result;
-    await axios.get('/rideEstimate', {
-      params: {
-        start_lat: start.lat(),
-        start_lng: start.lng(),
-        end_lat: end.lat(),
-        end_lng: end.lng(),
-        ride_type: 'lyft'
-      }
-    })
-    .then(res => {result = res})
-    .catch(errors => console.log(errors))
-
-    console.log(result);
-    this.props.newMarker(end);
-    if (result.data) {
-      let primetimeString = result.data.cost_estimates[0].primetime_percentage;
-      let primtimePercentage = parseFloat(primetimeString) / 100.0;
-      let baseCost = result.data.cost_estimates[0].estimated_cost_cents_max / 100;
-      let estimate = (primtimePercentage * baseCost) + baseCost;
-
-      // let estimate = result.data.cost_estimates[0].estimated_cost_cents_max / 100;
-      if ((estimate < (amount + stdDev) && estimate > (amount - stdDev)) ||
-          history.length > 15) {
-        let newBoundaries = Object.assign({}, this.state.boundaries);
-        newBoundaries[index] = end;
-        this.setState({ boundaries: newBoundaries },
-        () => {
-          if (Object.keys(this.state.boundaries).length === 18) {
-            // MapTools.drawBoundaries(this.state.boundaries, this.props.map);
-            this.props.drawBoundaries(this.state.boundaries);
-          }
-        });
-      } else {
-        let ratio = amount / estimate;
-        const googleGeometry = google.maps.geometry.spherical;
-        const newDistance = googleGeometry.computeDistanceBetween(start, end);
-        const newEnd = new googleGeometry.computeOffset(start, ratio * newDistance, direction);
-        history.push(newEnd);
-        this.rideEstimate(start, newEnd, amount, stdDev, index, direction, history);
-      }
-    }
+    this.setState({ rideType: type }, () => { this.getBoundaries(); })
   }
 
   render() {
@@ -227,7 +113,6 @@ class UserInputForm extends React.Component {
       </React.Fragment>
     );
   }
-
 }
 
 export default UserInputForm;
