+++ client/components/map/map_style.js
@@ -1,30 +1,3 @@
-// const mapStyle = [
-//   {
-//     featureType: 'road',
-//     elementType: 'all',
-//     stylers: [
-//       { visibility: 'on' },
-//       { lightness: 100 },
-//       { gamma: 9.99 }
-//     ]
-//   },
-//   {
-//     featureType: 'road',
-//     elementType: 'labels',
-//     stylers: [
-//       { visibility: 'off' }
-//     ]
-//   },
-//   {
-//     featureType: 'poi',
-//     elementType: 'labels',
-//     stylers: [
-//       { visibility: 'off' }
-//     ]
-//   }
-// ];
-
-
 const mapStyle = [
   {
     "elementType": "geometry",
