+++ client/util/latlong_conversion.js
@@ -1,15 +1,15 @@
 export const parseAddressToLatLng = function(address, callback) {
- const geocoder = new google.maps.Geocoder;
- geocoder.geocode({ address: address }, (results, status) => {
-   if (status === 'OK') {
-     const addressLatLng = new google.maps.LatLng(
-       results[0].geometry.location.lat(),
-       results[0].geometry.location.lng()
-     );
-     this.setState({ addressLatLng }, () => {
-       this.getBoundaries();
-     });
-     this.centerMap(addressLatLng);
-   }
- });
+  const geocoder = new google.maps.Geocoder;
+  geocoder.geocode({ address: address }, (results, status) => {
+    if (status === 'OK') {
+      const addressLatLng = new google.maps.LatLng(
+        results[0].geometry.location.lat(),
+        results[0].geometry.location.lng()
+      );
+      this.setState({ addressLatLng }, () => {
+        this.getBoundaries();
+      });
+      this.centerMap(addressLatLng);
+    }
+  });
 };
